Refactor camera types for docs purposes
