+++ index.ts
@@ -8,7 +8,6 @@ import type {
 } from 'react-native';
 import { NativeModules, PermissionsAndroid } from 'react-native';
 import type { Units, BBox, Id } from '@turf/helpers';
-import type { CameraAnimationMode, UserTrackingMode } from 'types/Camera';
 
 import {
   MapView,
@@ -36,7 +35,12 @@ import {
   MarkerView,
   Style,
 } from './javascript/components';
-import Camera, { CameraProps, CameraRef } from './javascript/components/Camera';
+import Camera, {
+  CameraAnimationMode,
+  CameraProps,
+  CameraRef,
+  UserTrackingMode,
+} from './javascript/components/Camera';
 import locationManager from './javascript/modules/location/locationManager';
 import offlineManager from './javascript/modules/offline/offlineManager';
 import snapshotManager from './javascript/modules/snapshot/snapshotManager';
