--- types/Camera.ts
@@ -1,20 +1,0 @@
-import type { MapboxGLEvent } from 'index';
-
-export type CameraAnimationMode =
-  | 'flyTo'
-  | 'easeTo'
-  | 'linearTo'
-  | 'moveTo'
-  | 'none';
-
-export type UserTrackingMode = 'normal' | 'compass' | 'course';
-
-export type UserTrackingModeChangeCallback = (
-  event: MapboxGLEvent<
-    'usertrackingmodechange',
-    {
-      followUserLocation: boolean;
-      followUserMode: UserTrackingMode | null;
-    }
-  >,
-) => void;
